+++ api_docs/changelog.md
@@ -33,8 +33,12 @@ format used by the Zulip server that they are interacting with.
 * [`POST /register`](/api/register-queue): Added
   `max_channel_folder_name_length` and
   `max_channel_folder_description_length` fields to the response.
-- Mobile push notification payloads for APNs no longer contain the
-  `time` field, which was unused.
+* Mobile push notification payloads for APNs no longer contain the
+  `time`, `server`, and `realm_id` fields, which were unused.
+* Mobile push notification payloads for FCM to remove push notifications
+  no longer contain the `zulip_message_id` field. Clients that use `zulip_message_ids`
+  have been out for years and we no longer support older client versions
+  that don't.
 
 **Feature level 409**